get_extended_budget

Get the flow rate across cell faces including potential stresses applied along boundaries at a given time. Only implemented for "classical" MODFLOW versions where the budget is recorded as FLOW RIGHT FACE, FLOW FRONT FACE and FLOW LOWER FACE arrays.

get_extended_budget(cbcfile, ...)

Arguments

cbcfile
Cell by cell file produced by Modflow.

...
additional arguments passed to flopy$utils$postprocessing$get_extended_budget for information see references below

Value

return sub-list for with "Qx_ext", "Qy_ext", "Qz_ext" for each budget output timestep. Flow rates across cell faces. Qx_ext is a array of size (nlay, nrow, ncol + 1). Qy_ext is a array of size (nlay, nrow + 1, ncol). The sign is such that the y axis is considered to increase in the north direction. Qz_ext is a ndarray of size (nlay + 1, nrow, ncol). The sign is such that the z axis is considered to increase in the upward direction.

References

https://flopy.readthedocs.io/en/latest/source/flopy.utils.postprocessing.html#flopy.utils.postprocessing.get_extended_budget

Examples

 if (FALSE) {
 flopy <- import_flopy()
 reticulate::py_help(object = flopy$utils$postprocessing$get_extended_budget)
 }